Add drop none method
