+++ src/dgipy/graph_app.py
@@ -6,7 +6,7 @@ from dash import Input, Output, State, ctx, dash, dcc, html
 
 from dgipy import dgidb
 from dgipy import network_graph as ng
-from dgipy.type_utils import make_tabular
+from dgipy.data_utils import make_tabular
 
 
 def generate_app() -> dash.Dash:
